drivers: misc: pio_rpi_pico: Make `pio_rpi_pico_get_pio` inline
As this short function is frequently invoked from every PIO-based
drivers to get the PIO structure.
